+++ include/Vector2D.h
@@ -3,10 +3,10 @@
 class Vector2D
 {
 public:
-    int x, y;
+    float x, y;
 
     Vector2D();
-    Vector2D(int x, int y);
+    Vector2D(float x, float y);
 
     Vector2D& Add(const Vector2D& vec);
 	Vector2D& Subtract(const Vector2D& vec);
@@ -19,7 +19,7 @@ public:
     Vector2D& operator/=(const Vector2D& vec);
 
     Vector2D& Perp();
-    int Dot(const Vector2D& vec);
+    float Dot(const Vector2D& vec);
 
     Vector2D& Normalize();
